linting: intentify-sol

+++ contracts/intentify/src/WalletFactory.sol
@@ -6,7 +6,15 @@ import { SafeProxy } from "safe-contracts/proxies/SafeProxy.sol";
 import { SafeProxyFactory } from "safe-contracts/proxies/SafeProxyFactory.sol";
 
 contract WalletFactory is SafeProxyFactory {
-    function getDeterministicWalletAddress(address _singleton, address owner, uint256 saltNonce) public view returns (address proxy) {
+    function getDeterministicWalletAddress(
+        address _singleton,
+        address owner,
+        uint256 saltNonce
+    )
+        public
+        view
+        returns (address proxy)
+    {
         address[] memory owners = new address[](1);
         owners[0] = owner;
         bytes memory initializer = abi.encodeWithSelector(
@@ -22,14 +30,19 @@ contract WalletFactory is SafeProxyFactory {
         );
         bytes32 salt = keccak256(abi.encodePacked(keccak256(initializer), saltNonce));
         bytes memory deploymentData = abi.encodePacked(type(SafeProxy).creationCode, uint256(uint160(_singleton)));
-        bytes32 hash = keccak256(
-            abi.encodePacked(bytes1(0xff), address(this), salt, keccak256(deploymentData))
-        );
+        bytes32 hash = keccak256(abi.encodePacked(bytes1(0xff), address(this), salt, keccak256(deploymentData)));
         // NOTE: cast last 20 bytes of hash to address
-        return address(uint160(uint(hash)));
+        return address(uint160(uint256(hash)));
     }
 
-    function createDeterministicWallet(address _singleton, address owner, uint256 saltNonce) public returns (SafeProxy proxy) {
+    function createDeterministicWallet(
+        address _singleton,
+        address owner,
+        uint256 saltNonce
+    )
+        public
+        returns (SafeProxy proxy)
+    {
         bytes memory initializer = _encodeInitializer(owner);
         bytes32 salt = keccak256(abi.encodePacked(keccak256(initializer), saltNonce));
         proxy = deployProxy(_singleton, initializer, salt);
